Fix: 구독취소 예외처리추가(구독 없을경우)

+++ src/channel/channel.service.ts
@@ -49,16 +49,9 @@ export class ChannelService {
   }
 
   async createSubscribe(userIdx: number, channelIdx: number): Promise<void> {
-    console.log(userIdx);
-    console.log(channelIdx);
+    const subscribeState = await this.getSubscrbeState(userIdx, channelIdx);
 
-    const existingSubscribe = await this.prisma.subscribe.findUnique({
-      where: {
-        subscriber_provider: { subscriber: userIdx, provider: channelIdx },
-      },
-    });
-
-    if (existingSubscribe) {
+    if (subscribeState) {
       throw new ConflictException('alreay subscribe');
     }
 
@@ -71,6 +64,12 @@ export class ChannelService {
     userIdx: number,
     channelIdx: number,
   ): Promise<void> {
+    const subscribeState = await this.getSubscrbeState(userIdx, channelIdx);
+
+    if (!subscribeState) {
+      throw new ConflictException('alreay subscribe');
+    }
+
     await this.prisma.subscribe.deleteMany({
       where: { subscriber: userIdx, provider: channelIdx },
     });
